fix unittest
