+++ test/prim/prim/vjp/static/test_comp_reshape_grad.py
@@ -105,28 +105,9 @@ class TestReshapeGradComp(unittest.TestCase):
 
         return res
 
-    def test_cinn(self):
-        paddle.disable_static()
-        use_cinn = True
-        if isinstance(
-            framework._current_expected_place(), framework.core.CPUPlace
-        ):
-            # TODO(jiabin): CINN will crashed in this case open it when fixed
-            use_cinn = False
-
-        dy_res = self.train(use_prim=False, use_cinn=False)
-        comp_st_cinn_res = self.train(use_prim=True, use_cinn=use_cinn)
-
-        for i in range(len(dy_res)):
-            np.testing.assert_allclose(
-                comp_st_cinn_res[i].numpy(),
-                dy_res[i].numpy(),
-                rtol=1e-7,
-                atol=1e-7,
-            )
+    def test_reshape_grad_comp(self):
         paddle.enable_static()
 
-    def test_reshape_grad_comp(self):
         def actual(primal, shape, cotangent):
             core._set_prim_backward_enabled(True)
             mp, sp = paddle.static.Program(), paddle.static.Program()
@@ -143,7 +124,7 @@ class TestReshapeGradComp(unittest.TestCase):
             return exe.run(
                 program=mp,
                 feed={'primal': primal, 'cotangent': cotangent},
-                fetch_list=[x_cotangent[0].name],
+                fetch_list=[x_cotangent[0]],
             )[0]
 
         def desired(primal, shape, cotangent):
@@ -162,7 +143,7 @@ class TestReshapeGradComp(unittest.TestCase):
             return exe.run(
                 program=mp,
                 feed={'primal': primal, 'cotangent': cotangent},
-                fetch_list=[x_cotangent[0].name],
+                fetch_list=[x_cotangent[0]],
             )[0]
 
         if (self.dtype == np.float16) and isinstance(
@@ -178,6 +159,7 @@ class TestReshapeGradComp(unittest.TestCase):
                 atol=self.rtol,
             )
         core._set_prim_backward_enabled(False)
+        paddle.disable_static()
 
 
 if __name__ == '__main__':
